+++ src/mcts_node.h
@@ -26,6 +26,6 @@ Square getMostPromisingMove(MCTSNode* node, Board* board);
 
 int getSims(MCTSNode* node);
 
-double getWinrate(MCTSNode* node);
+float getWinrate(MCTSNode* node);
 
 #endif //UTTT2_MCTS_NODE_H
